[xss] Refactor #839
