+++ lib/models/law.js
@@ -12,7 +12,7 @@ require('node-path')(module);
 var mongoose = require('mongoose');
 var validators = require('mongoose-validators')
 var log = require('debug')('democracyos:models:law');
-var cleanSummary = require('./lib/clean-summary');
+var xss = require('lib/richtext').xssFilter({ video: true, image: true, link: true });
 
 var Schema = mongoose.Schema;
 var ObjectId = Schema.ObjectId;
@@ -133,7 +133,7 @@ LawSchema.options.toJSON.transform = function(doc, ret) {
 
 LawSchema.pre('save', function(next) {
   this.updatedAt = this.isNew ? this.createdAt : Date.now();
-  this.summary = cleanSummary(this.summary);
+  this.summary = xss(this.summary);
 
   next();
 });
